backend/requirements.txt
````txt
Flask
Flask-Cors
SQLAlchemy
psycopg2-binary
python-dotenv
bcrypt
PyJWT
cryptography
requests
alembic
````